- custom base64 encoder removed; - warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@5117 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/WhatsApp/src/WhatsAPI++/ByteArray.cpp
@@ -5,6 +5,7 @@
  *      Author: Antonio
  */
 
+#include "../common.h"
 #include "ByteArray.h"
 #include "WAException.h"
 #include <iostream>
@@ -116,7 +117,7 @@ int ByteArrayInputStream::read(std::vector<unsigned char>& b, size_t  off, size_
 		}
 	} catch (std::exception& ) {
 	}
-	return i;
+	return (int)i;
 }
 
 ByteArrayInputStream::~ByteArrayInputStream() {
